pyproject.toml
[project]
name = "mcphttpserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "faiss-cpu>=1.11.0",
    "mcp[cli]>=1.9.4",
    "numpy>=2.3.1",
    "openai>=1.93.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "tiktoken>=0.9.0",
]
